Profile apis integration

Approved-by: Aftab Bhadki

+++ networklayer/src/main/java/com/emproto/networklayer/response/responsee/ValueXX.kt
@@ -3,7 +3,7 @@ package com.emproto.networklayer.response.responsee
 data class ValueXX(
     val height: Int,
     val mediaType: String,
-    val size: Double,
+    val size: Int,
     val url: String,
     val width: Int
 )
